requirements.txt
torch>=2.2.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.6
datasets>=2.19.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
wandb>=0.16.0
rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.1
streamlit>=1.33.0
pyyaml>=6.0
numpy>=1.26.0
scipy>=1.11.0